fix: Namespaced instructionStep
