+++ app/graphql/types/tfl/entities/instruction_type.rb
@@ -3,7 +3,7 @@ module Types
     module Entities
       class InstructionType < Types::BaseObject
         field :detailed, String, null: true
-        field :steps, [Types::InstructionStepType, null: true], null: true
+        field :steps, [Types::Tfl::Entities::InstructionStepType, null: true], null: true
         field :summary, String, null: true
       end
     end
